Add files via upload
In this image, it shows our Bank Transaction expression.

The console greets the user, "lecome to DDK bank..." and asks how many transactions the user wants to execute. The user puts 2, and the console runs the transaction twice, asking if it wants to withdraw or deposit each time, each time subtracting/adding to the balance (99000). After two transactions have been executed, console will ask the user if it wants to make another transaction to which the user puts y (y to repeat). It executes another transaction, however it goes back to the stated balance of 99000 (We'll hopefully learn how to avoid that.) and deposits/withdraws from that balance, then it gives the user a message thanking them for using DDK bank :).